% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arabidopsis}
\alias{arabidopsis}
\title{Arabidopsis cytosine data}
\format{A data frame with 8 rows and 3 columns
\itemize{
  \item CG: Number of cytosines in the CG context
  \item CHG: Number of cytosines in the CHG context
  \item CHH: Number of cytosines in the CHH context
}}
\source{
\url{ftp://igenome:G3nom3s4u@ussd-ftp.illumina.com/Arabidopsis_thaliana/Ensembl/TAIR10/Arabidopsis_thaliana_Ensembl_TAIR10.tar.gz}
}
\usage{
arabidopsis
}
\description{
A dataframe containing cytosine counts for each sequence context and chromosome
}
\keyword{datasets}

