% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{check_sample_metadata}
\alias{check_sample_metadata}
\title{Cross-check reported metadata against computed QC metrics (logging only)}
\usage{
check_sample_metadata(sample_sheet, qc_df, logger = NULL)
}
\description{
Cross-check reported metadata against computed QC metrics (logging only)
}
