% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{compute_qc_metrics}
\alias{compute_qc_metrics}
\title{Compute per-sample QC metrics from a list of SigDFs}
\usage{
compute_qc_metrics(sdfs, platform, logger = NULL)
}
\description{
Compute per-sample QC metrics from a list of SigDFs
}
