% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageSurvival.R
\name{coverageSurvival}
\alias{coverageSurvival}
\title{Coverage Survival}
\usage{
coverageSurvival(data, cytosines, chromosome = "all")
}
\arguments{
\item{data}{A dataframe}

\item{cytosines}{Dataframe with cytosine counts. Can be generated using methylQC::cytosines()}

\item{chromosome}{Chromosome number. Defaults to all (all chromosomes)}
}
\value{
a dataframe
}
\description{
Find the diminishing percentage of cytosines with increasing sequencing depth
}
\examples{
coverage <- coverageSurvival(data = methylome, cytosines = arabidopsis, chromosome = "chr1")
}

