% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosines.R
\name{cytosines}
\alias{cytosines}
\title{Cytosines}
\usage{
cytosines(data)
}
\arguments{
\item{data}{Path to fasta file}
}
\value{
a dataframe
}
\description{
Count the number of cytosines in each sequence context, for each chromosome
}
\examples{
path <- system.file("extdata", "lambda.fasta", package = "methylQC")
cytosines(path)
}

