% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{edit_exclusions}
\alias{edit_exclusions}
\title{Edit exclusion lists post-hoc}
\usage{
edit_exclusions(
  stage1_dir,
  add_samples = character(0),
  remove_samples = character(0),
  add_probes = character(0),
  remove_probes = character(0),
  add_reason = "manual"
)
}
\description{
Edit exclusion lists post-hoc
}
