% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(path, datasource = "BSseeker2")
}
\arguments{
\item{path}{Path to the data}

\item{datasource}{Which bisulfite package was used? Defaults to BSseeker2}
}
\value{
a dataframe
}
\description{
This function will load the dataset to be used in methylQC functions
}
\examples{
loadData("~/Documents/mC_data.CGmap.gz")
}
\keyword{data}

