% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_capture}
\alias{log_capture}
\title{Append a captured object to the logger's file}
\usage{
log_capture(logger, obj)
}
\arguments{
\item{logger}{A logger from \code{\link{make_logger}}.}

\item{obj}{Any object; printed via \code{print()}.}
}
\description{
Append a captured object to the logger's file
}
