% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logcapture}
\alias{logcapture}
\title{Append a captured object to the logger's file}
\usage{
logcapture(logger, obj)
}
\arguments{
\item{logger}{A logger from \code{\link{makelog}}.}

\item{obj}{Any object; printed via \code{print()}.}
}
\description{
Append a captured object to the logger's file
}
