% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{makelog}
\alias{makelog}
\title{Create a file + console logger}
\usage{
makelog(outdir)
}
\arguments{
\item{outdir}{Directory where the log file will be written.}
}
\value{
A list with \code{log(step, msg)} function and \code{path}.
}
\description{
Create a file + console logger
}
