% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_txt.R
\name{methodstext}
\alias{methodstext}
\title{Compose the Methods text}
\usage{
methodstext(i)
}
\arguments{
\item{i}{a list of run facts.}
}
\value{
a character vector of lines.
}
\description{
Compose the Methods text
}
\examples{
methodstext(list(platform = "EPIC", n_samples = 10L, n_probes = 866553L))
}
