% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylomeStats.R
\name{methylomeStats}
\alias{methylomeStats}
\title{Methylome statistics}
\usage{
methylomeStats(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
a dataframe
}
\description{
Generate summary statistics for methylome data
}
\examples{
methylomeStats(dat)
}

