% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{methylQC_options}
\alias{methylQC_options}
\title{Get current methylQC configuration}
\usage{
methylQC_options(...)
}
\arguments{
\item{...}{Named overrides to apply temporarily.}
}
\value{
A named list of all current options.
}
\description{
Get current methylQC configuration
}
