% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{methylQC_set}
\alias{methylQC_set}
\title{Set methylQC options globally}
\usage{
methylQC_set(...)
}
\arguments{
\item{...}{Named options (e.g., \code{donor_col = "SubjectID"}).}
}
\value{
Invisibly, the previous values.
}
\description{
Set methylQC options globally
}
