% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mqcopts}
\alias{mqcopts}
\title{Get current methylQC configuration}
\usage{
mqcopts(...)
}
\arguments{
\item{...}{Named overrides to apply temporarily to the returned list.}
}
\value{
A named list of all current options.
}
\description{
Get current methylQC configuration
}
