% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mqcreset}
\alias{mqcreset}
\title{Reset methylQC options to their defaults}
\usage{
mqcreset()
}
\value{
\code{NULL}, invisibly.
}
\description{
Reset methylQC options to their defaults
}
\examples{
mqcreset()
}
