% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mqcset}
\alias{mqcset}
\title{Set methylQC options globally}
\usage{
mqcset(...)
}
\arguments{
\item{...}{Named options (e.g., \code{donorcol = "SubjectID"}).}
}
\value{
Invisibly, the previous values.
}
\description{
Set methylQC options globally
}
