% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonConversion.R
\name{nonConversion}
\alias{nonConversion}
\title{Non-conversion}
\usage{
nonConversion(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
a dataframe
}
\description{
Calculate the bisulfite non-conversion rate
}
\examples{
nonConversion(methylome)
}

