% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrowser.R
\name{plotBrowser}
\alias{plotBrowser}
\title{Plot Browser}
\usage{
plotBrowser(dat, start = 5000, stop = 15000)
}
\arguments{
\item{start}{The start position. Defaults to 5000}

\item{data}{A dataframe}

\item{start}{The stop position. Defaults to 15000}
}
\value{
ggplot2 object
}
\description{
This function will plot a browser view of sequencing coverage
}
\examples{
plotBrowser(dat)
}

