% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvival.R
\name{plotSurvival}
\alias{plotSurvival}
\title{Plot Survival}
\usage{
plotSurvival(data, species = "Arabidopsis", chromosome = "all")
}
\arguments{
\item{data}{A dataframe}

\item{species}{Species name. Defaults to Arabidopsis}

\item{chromosome}{Chromosome number. Defaults to all (all chromosomes)}
}
\value{
ggplot2 plot object
}
\description{
Plot the diminishing percentage of cytosines with increasing sequencing depth
}
\examples{
plotBrowser(dat)
}

