% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{prep_single}
\alias{prep_single}
\title{Stage 2: Flagging, QC report, EpiDISH, consolidated sample sheet}
\usage{
prep_single(
  cell_type,
  betas_sub,
  mvals_sub,
  ss_sub,
  platform,
  out_subdir,
  mask_sub = NULL,
  detP_sub = NULL,
  ...
)
}
\arguments{
\item{mask_sub}{Logical matrix (probes x samples). NULL if not available.}

\item{detP_sub}{Numeric matrix of detection p-values. NULL if not available.}
}
\description{
Stage 2: Flagging, QC report, EpiDISH, consolidated sample sheet
}
