% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{procmem}
\alias{procmem}
\title{Resident memory of the current R process}
\usage{
procmem()
}
\value{
bytes, or \code{NA_real_} if unavailable.
}
\description{
Resident memory of the current R process
}
\examples{
procmem()
}
