% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandBias.R
\name{strandBias}
\alias{strandBias}
\title{Strand sequencing bias}
\usage{
strandBias(data, window = 30, fill = 0)
}
\arguments{
\item{data}{input dataframe}

\item{window}{size of sliding window. Default is 30 bases.}

\item{fill}{Value used to fill in where there isn't enough data for window size. Default is 0.}
}
\value{
a dataframe
}
\description{
This function will calculate the strand depth bias for each C base position
}
\examples{
strandBias(data)
}

