% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_qc_report}
\alias{write_qc_report}
\title{Write all QC diagnostic plots to a single PDF}
\usage{
write_qc_report(
  ss_all,
  betas_all_sub,
  betas_qc,
  excluded_ids,
  platform,
  pdf_path,
  pc_csv_path,
  probe_csv_path,
  n_top = 10000,
  logger = NULL
)
}
\description{
Write all QC diagnostic plots to a single PDF
}
