% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_txt.R
\name{writemethods}
\alias{writemethods}
\title{Write the run's Methods description}
\usage{
writemethods(dir, info)
}
\arguments{
\item{dir}{output directory.}

\item{info}{a list of run facts assembled by \code{\link{qc}}.}
}
\value{
the path written, invisibly.
}
\description{
Write the run's Methods description
}
\examples{
\dontrun{
writemethods("~/qcout", info)
}
}
