% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{compute_qc_metrics_streaming}
\alias{compute_qc_metrics_streaming}
\title{Streaming QC metrics}
\usage{
compute_qc_metrics_streaming(basenames, platform, logger = NULL)
}
\description{
Streaming QC metrics
}
