% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logcapture}
\alias{logcapture}
\title{Capture output, messages and warnings from an expression into a log}
\usage{
logcapture(expr, logger, stage = "capture")
}
\arguments{
\item{expr}{expression to evaluate}

\item{logger}{a logger from \code{\link{makelog}}}

\item{stage}{stage label recorded against each captured line}
}
\value{
the value of \code{expr}
}
\description{
Capture output, messages and warnings from an expression into a log
}
\examples{
lg <- makelog(file.path(tempdir(), "demo2.log"))
logcapture(message("working"), lg, "demo")
lg$close()
}
