% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mqcdefaults}
\alias{mqcdefaults}
\title{Default methylQC options}
\usage{
mqcdefaults()
}
\value{
a named list of the factory defaults.
}
\description{
Default methylQC options
}
\examples{
mqcdefaults()$detp
}
