% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mqcopts}
\alias{mqcopts}
\title{Get current methylQC options}
\usage{
mqcopts(...)
}
\arguments{
\item{...}{optional option names; if given, only those are returned.}
}
\value{
a named list of the currently active options.
}
\description{
Get current methylQC options
}
\examples{
mqcopts("detp", "maskuse")
}
