% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Plot coverage}
\usage{
plotCoverage(data)
}
\arguments{
\item{data}{A dataframe. Can be generated using methylQC::coverageSurvival()}
}
\value{
a ggplot2 object
}
\description{
Plot the diminishing percentage of cytosines with increasing sequencing depth
}
\examples{
coverage <- coverageSurvival(data = methylome, cytosines = arabidopsis, chromosome = "L")
plotCoverage(coverage)
}

