% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvival.R
\name{plotSurvival}
\alias{plotSurvival}
\title{Plot Survival}
\usage{
plotSurvival(data, cytosines, chromosome = "all")
}
\arguments{
\item{data}{A dataframe}

\item{cytosines}{Dataframe with cytosine counts. Can be generated using methylQC::cytosines()}

\item{chromosome}{Chromosome number. Defaults to all (all chromosomes)}
}
\value{
ggplot2 plot object
}
\description{
Plot the diminishing percentage of cytosines with increasing sequencing depth
}
\examples{
data <- loadData("methylome.CGmap")
cytosines <- cytosines("genome.fa")
plotSurvival(data, cytosines, chromosome = "chr1")
}

