% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{prep}
\alias{prep}
\title{Stage 2 wrapper: optionally split by cell type}
\usage{
prep(stage1_dir, platform = NULL, split_by_cell = TRUE, ...)
}
\description{
Stage 2 wrapper: optionally split by cell type
}
