% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidish.R
\name{run_epidish}
\alias{run_epidish}
\title{Run EpiDISH deconvolution on a beta matrix}
\usage{
run_epidish(betas, reference = NULL, method = NULL, logger = NULL)
}
\description{
Run EpiDISH deconvolution on a beta matrix
}
